refactor: move middlewares into core module
