+++ central-oon-core-backend/index.js
@@ -1,7 +1,9 @@
 const { createApp } = require("./app");
 const { startServer } = require("./server");
+const middlewares = require("./middlewares");
 
 module.exports = {
   createApp,
   startServer,
+  middlewares,
 };
