+++ src/routers/assistenteRouter.js
@@ -2,8 +2,8 @@ const express = require("express");
 const router = express.Router();
 const AssistenteController = require("../controllers/assistente");
 const {
-  registrarAcaoMiddleware,
-} = require("../middlewares/registrarAcaoMiddleware");
+  middlewares: { registrarAcaoMiddleware },
+} = require("../../central-oon-core-backend");
 const { ACOES, ENTIDADES } = require("../constants/controleAlteracao");
 const { asyncHandler } = require("../utils/helpers");
