+++ src/routers/documentoFiscalRouter.js
@@ -5,8 +5,8 @@ const router = express.Router();
 
 const { uploadExcel, uploadPDFAndImage } = require("../config/multer");
 const {
-  registrarAcaoMiddleware,
-} = require("../middlewares/registrarAcaoMiddleware");
+  middlewares: { registrarAcaoMiddleware },
+} = require("../../central-oon-core-backend");
 const { ACOES, ENTIDADES } = require("../constants/controleAlteracao");
 const { asyncHandler } = require("../utils/helpers");
