+++ src/routers/etapaRouter.js
@@ -2,8 +2,8 @@ const express = require("express");
 const router = express.Router();
 const EtapaController = require("../controllers/etapa");
 const {
-  registrarAcaoMiddleware,
-} = require("../middlewares/registrarAcaoMiddleware");
+  middlewares: { registrarAcaoMiddleware },
+} = require("../../central-oon-core-backend");
 const { ACOES, ENTIDADES } = require("../constants/controleAlteracao");
 const { asyncHandler } = require("../utils/helpers");
