+++ src/routers/integracaoRouter.js
@@ -2,8 +2,8 @@ const express = require("express");
 const IntegracaoController = require("../controllers/integracao");
 const { asyncHandler } = require("../utils/helpers");
 const {
-  registrarAcaoMiddleware,
-} = require("../middlewares/registrarAcaoMiddleware");
+  middlewares: { registrarAcaoMiddleware },
+} = require("../../central-oon-core-backend");
 const { ACOES, ENTIDADES } = require("../constants/controleAlteracao");
 
 const router = express.Router();
