+++ src/routers/listasOmieRouter.js
@@ -3,8 +3,8 @@ const router = express.Router();
 const { ListaOmieController } = require("../controllers/listaOmieController");
 
 const {
-  registrarAcaoMiddleware,
-} = require("../middlewares/registrarAcaoMiddleware");
+  middlewares: { registrarAcaoMiddleware },
+} = require("../../central-oon-core-backend");
 const { ACOES, ENTIDADES } = require("../constants/controleAlteracao");
 const { asyncHandler } = require("../utils/helpers");
