+++ src/routers/usuarioRouter.js
@@ -1,8 +1,8 @@
 const express = require("express");
 const UsuarioController = require("../controllers/usuario");
 const {
-  registrarAcaoMiddleware,
-} = require("../middlewares/registrarAcaoMiddleware");
+  middlewares: { registrarAcaoMiddleware },
+} = require("../../central-oon-core-backend");
 const { ACOES, ENTIDADES } = require("../constants/controleAlteracao");
 const router = express.Router();
 const { asyncHandler } = require("../utils/helpers");
